+++ src/main/java/io/beanmapper/annotations/BeanRecordConstructMode.java
@@ -4,9 +4,9 @@ package io.beanmapper.annotations;
  * An enum used to set whether a constructor of a Record should be considered for mapping, force to be used for mapping,
  * or excluded from mapping.
  *
- * @see RecordConstruct
+ * @see BeanRecordConstruct
  */
-public enum RecordConstructMode {
+public enum BeanRecordConstructMode {
 
     ON_DEMAND,
     FORCE,