+++ src/main/java/io/beanmapper/config/FlushAfterClearInstruction.java
@@ -1,7 +1,0 @@
-package io.beanmapper.config;
-
-public enum FlushAfterClearInstruction {
-
-    FLUSH_ENABLED, FLUSH_DISABLED, UNSET
-
-}