+++ src/main/java/io/beanmapper/config/Trinary.java
@@ -1,0 +1,7 @@
+package io.beanmapper.config;
+
+public enum Trinary {
+
+    ENABLED, DISABLED, UNSET
+
+}