+++ src/main/java/io/beanmapper/core/BeanMatchStore.java
@@ -24,7 +24,7 @@ import io.beanmapper.annotations.BeanRoleSecured;
 import io.beanmapper.annotations.BeanUnwrap;
 import io.beanmapper.config.BeanPair;
 import io.beanmapper.config.CollectionHandlerStore;
-import io.beanmapper.config.FlushAfterClearInstruction;
+import io.beanmapper.config.Trinary;
 import io.beanmapper.core.collections.CollectionHandler;
 import io.beanmapper.core.converter.collections.AnnotationClass;
 import io.beanmapper.core.converter.collections.BeanCollectionInstructions;
@@ -225,7 +225,7 @@ public class BeanMatchStore {
         CollectionElementType elementType = EMPTY_COLLECTION_ELEMENT_TYPE;
         BeanCollectionUsage beanCollectionUsage = null;
         AnnotationClass preferredCollectionClass = EMPTY_ANNOTATION_CLASS;
-        FlushAfterClearInstruction flushAfterClear = FlushAfterClearInstruction.UNSET;
+        Trinary flushAfterClear = Trinary.UNSET;
 
         CollectionHandler collectionHandler = null;
         if (beanCollection == null) {