+++ src/test/java/io/beanmapper/strategy/record/model/annotation/record_construct/RecordWithAllConstructorsExcluded.java
@@ -1,11 +1,11 @@
 package io.beanmapper.strategy.record.model.annotation.record_construct;
 
-import io.beanmapper.annotations.RecordConstruct;
-import io.beanmapper.annotations.RecordConstructMode;
+import io.beanmapper.annotations.BeanRecordConstruct;
+import io.beanmapper.annotations.BeanRecordConstructMode;
 
 public record RecordWithAllConstructorsExcluded(int id, String name) {
 
-    @RecordConstruct(constructMode = RecordConstructMode.EXCLUDE)
+    @BeanRecordConstruct(constructMode = BeanRecordConstructMode.EXCLUDE)
     public RecordWithAllConstructorsExcluded {
     }