+++ src/test/java/io/beanmapper/testmodel/record/TargetRecordWithMultipleForcedConstructors.java
@@ -1,16 +1,16 @@
 package io.beanmapper.testmodel.record;
 
-import io.beanmapper.annotations.RecordConstruct;
-import io.beanmapper.annotations.RecordConstructMode;
+import io.beanmapper.annotations.BeanRecordConstruct;
+import io.beanmapper.annotations.BeanRecordConstructMode;
 
 public record TargetRecordWithMultipleForcedConstructors(int id, String name) {
 
-    @RecordConstruct(value = "id", constructMode = RecordConstructMode.FORCE)
+    @BeanRecordConstruct(value = "id", constructMode = BeanRecordConstructMode.FORCE)
     public TargetRecordWithMultipleForcedConstructors(int id) {
         this(id, "Henk");
     }
 
-    @RecordConstruct(value = "name", constructMode = RecordConstructMode.FORCE)
+    @BeanRecordConstruct(value = "name", constructMode = BeanRecordConstructMode.FORCE)
     public TargetRecordWithMultipleForcedConstructors(String name) {
         this(1, name);
     }