+++ CHANGELOG.md
@@ -29,7 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   Collection, will return an empty Collection of the target type, rather than returning null.
 - Issue [#153](https://github.com/42BV/beanmapper/issues/153) **https://github.com/42BV/beanmapper/issues/153**; Rather than using the Boolean-wrapper, all
   occurrences of Boolean that are not absolutely necessary due to generics, have been replaced with the primitive boolean, or the
-  FlushAfterClearInstruction-enum.
+  Trinary-enum.
 - Issue [#166](https://github.com/42BV/beanmapper/issues/166) **Source with BeanAlias-annotated fields cannot be downsized.**; Modified
   ClassGenerator#createClass, to check whether a generated field is annotated with BeanAlias. If so, the name of the generated field will be set to the value on
   the BeanAlias-annotation, and the annotation will be removed from the generated field.
@@ -44,8 +44,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - BeanMapper#map(Queue, Class) allowing users to map a Queue to a new Queue, mapping the elements of the source to the target class.
 - BeanMapper#map(Object[], Class) allowing users to map an array to an array with elements of the type of the target class. Also works for primitive arrays.
 - Support for mapping to and from JDK16 record-classes.
-- @RecordConstruct-annotation, used to give BeanMapper instructions on how to map a record.
-- RecordConstructMode-enum, which allows the user to exclude certain constructors from being used to map a record, or force one to be used.
+- @BeanRecordConstruct-annotation, used to give BeanMapper instructions on how to map a record.
+- BeanRecordConstructMode-enum, which allows the user to exclude certain constructors from being used to map a record, or force one to be used.
 - Configuration#addCustomDefaultValueForClass(Class<?>, Object value) and Configuration#getDefaultValueForClass(Class<?>), which can be used to define default
   values for classes during runtime.
 - RecordToAnyConverter, which creates a dynamic class based off of the given Record, which contains only public fields, corresponding to the RecordComponents.